requirements.txt
langchain-groq>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
